Currently a (filtered) site collection is a long array containing an attribute `.indices` from which shorter arrays are extracted. This extraction process is costly. Here I replace the filtered site collection with a shorter array. This change reduces a lot the distance calculations in`make_contexts`. This is the improvement for the South America calculation with 480,000 sites:

master
```
   ============================== ======== ========= ======
   operation                      time_sec memory_mb counts
   ============================== ======== ========= ======
   total classical                208,145  124       652   
   make_contexts                  130,843  0.0       93,656
   get_poes                       75,667   0.0       93,656
   total runtime                  1,746    1,110     1     
```
this branch
```
   ============================== ======== ========= ======
   operation                      time_sec memory_mb counts
   ============================== ======== ========= ======
   total classical                140,018  80        652   
   get_poes                       76,322   0.0       93,656
   make_contexts                  61,952   0.0       93,656
   total runtime                  1,320    1,262     1     
```
i.e. we more than doubled the speed of `make_contexts`. For the Central Africa calculation by Valerio the speedup is from 1,093,105 seconds to 537,451 seconds, still a factor of 2.